Add sorting for mtype to morphology endpoint
* Requested feature
* Had to further subclass fastapi_filter since it didn't take the
  sub-filters into account
