+++ tests/test_mtype.py
@@ -1,13 +1,6 @@
-from unittest.mock import ANY
-
-import pytest
-
-from app.db.model import MTypeClass, MTypeClassification
-
-from .utils import BEARER_TOKEN, PROJECT_HEADERS, add_db, create_reconstruction_morphology_id
+from app.db.model import MTypeClass
 
 ROUTE = "/mtype"
-ROUTE_MORPH = "/reconstruction-morphology"
 
 
 def test_mtype(db, client):
@@ -68,64 +61,3 @@ def test_missing_mtype(client):
 
     response = client.get(f"{ROUTE}/notanumber")
     assert response.status_code == 422
-
-
-@pytest.mark.usefixtures("skip_project_check")
-def test_morph_mtypes(db, client, species_id, strain_id, brain_region_id):
-    morph_id = create_reconstruction_morphology_id(
-        client,
-        species_id,
-        strain_id,
-        brain_region_id,
-        headers=BEARER_TOKEN | PROJECT_HEADERS,
-        authorized_public=False,
-    )
-
-    mtype1 = add_db(db, MTypeClass(pref_label="m1", alt_label="m1", definition="m1d"))
-    mtype2 = add_db(db, MTypeClass(pref_label="m2", alt_label="m2", definition="m2d"))
-
-    add_db(db, MTypeClassification(entity_id=morph_id, mtype_class_id=mtype1.id))
-    add_db(db, MTypeClassification(entity_id=morph_id, mtype_class_id=mtype2.id))
-
-    response = client.get(ROUTE_MORPH, headers=BEARER_TOKEN | PROJECT_HEADERS)
-    assert response.status_code == 200
-    facets = response.json()["facets"]
-    assert facets["mtype"] == [
-        {"id": 1, "label": "m1", "count": 1, "type": "mtype"},
-        {"id": 2, "label": "m2", "count": 1, "type": "mtype"},
-    ]
-
-    response = client.get(f"{ROUTE_MORPH}/{morph_id}", headers=BEARER_TOKEN | PROJECT_HEADERS)
-    assert response.status_code == 200
-    data = response.json()
-    assert "mtype" in data
-    mtype = data["mtype"]
-    assert len(mtype) == 2
-
-    assert mtype == [
-        {
-            "id": 1,
-            "pref_label": "m1",
-            "alt_label": "m1",
-            "definition": "m1d",
-            "creation_date": ANY,
-            "update_date": ANY,
-        },
-        {
-            "id": 2,
-            "pref_label": "m2",
-            "alt_label": "m2",
-            "definition": "m2d",
-            "creation_date": ANY,
-            "update_date": ANY,
-        },
-    ]
-
-    response = client.get(
-        f"{ROUTE_MORPH}?mtype__pref_label=m1", headers=BEARER_TOKEN | PROJECT_HEADERS
-    )
-    assert response.status_code == 200
-    facets = response.json()["facets"]
-    assert facets["mtype"] == [
-        {"id": 1, "label": "m1", "count": 1, "type": "mtype"},
-    ]
